refactor(domain): move enum classes and constants into domain package
