+++ src/main/kotlin/lotto/controller/LottoController.kt
@@ -3,9 +3,9 @@ package lotto.controller
 import lotto.Lotto
 import lotto.LottoManager
 import lotto.input.InputReader
-import lotto.input.end
-import lotto.input.start
-import lotto.input.ticketCost
+import lotto.domain.END_NUMBER
+import lotto.domain.START_NUMBER
+import lotto.domain.TICKET_COST
 import lotto.util.RandomNumbers
 import lotto.view.OutputView
 
@@ -14,10 +14,10 @@ object LottoController {
     fun run() {
 
         val purchaseAmount = InputReader.purchaseAmount()
-        val numberOfTickets = purchaseAmount / ticketCost
+        val numberOfTickets = purchaseAmount / TICKET_COST
 
         repeat(numberOfTickets) {
-            LottoManager.addTicket(Lotto(RandomNumbers.generate(start, end, 6)))
+            LottoManager.addTicket(Lotto(RandomNumbers.generate(START_NUMBER, END_NUMBER, 6)))
         }
 
         OutputView.printNumberOfTickets(numberOfTickets)
