+++ src/main/kotlin/lotto/domain/Constants.kt
@@ -1,0 +1,5 @@
+package lotto.domain
+
+const val TICKET_COST = 1000
+const val START_NUMBER = 1
+const val END_NUMBER = 45
