+++ src/test/kotlin/lotto/LottoManagerTest.kt
@@ -1,6 +1,7 @@
 package lotto
 
 import lotto.LottoManager.getTotalPrize
+import lotto.domain.Rank
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
