+++ src/test/kotlin/lotto/LottoTest.kt
@@ -1,5 +1,6 @@
 package lotto
 
+import lotto.domain.Rank
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
